rename to Chatmessages.js/css because they hold multiple components
